Migrate input from /docs/components/input to /components/input
Move all examples (Input Types, Disabled, Value Binding, Focus State)
from the old documentation page to the new reference page, remove the
old page, and update all navigation links and E2E tests.
